requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
Django==4.0.4
djangorestframework==3.13.1
fonttools==4.33.0
gensim==4.1.2
idna==3.3
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.2
matplotlib==3.5.1
neuralnet==0.2.0
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
requests==2.27.1
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
sqlparse==0.4.2
Theano==1.0.5
torch==1.11.0
torchfile==0.1.0
tornado==6.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
visdom==0.1.8.9
websocket-client==1.3.2
